split up file
